GroundwaterFlowProcess: move definitions to cpp
